Rename Parsers to Visitors and make their access modifier as internal
