--- VirtoCommerce.Storefront/Domain/StaticContent/ContentParsers/BlogMetadataParser.cs
@@ -1,46 +1,0 @@
-using System.Linq;
-using VirtoCommerce.Storefront.Model.StaticContent;
-
-namespace VirtoCommerce.Storefront.Domain
-{
-    public partial class StaticContentItemBuilder
-    {
-        private class BlogMetadataParser : IContentItemParser
-        {
-            public bool Suit(ContentItem item)
-            {
-                return item is BlogArticle;
-            }
-
-            public void Parse(string path, string content, ContentItem item)
-            {
-                var post = (BlogArticle)item;
-
-                if (post.MetaInfo.ContainsKey("main-image"))
-                {
-                    post.ImageUrl = post.MetaInfo["main-image"].FirstOrDefault();
-                }
-
-                if (post.MetaInfo.ContainsKey("excerpt"))
-                {
-                    post.Excerpt = post.MetaInfo["excerpt"].FirstOrDefault();
-                }
-
-                if (post.MetaInfo.ContainsKey("is-sticked"))
-                {
-                    bool.TryParse(post.MetaInfo["is-sticked"].FirstOrDefault(), out var isSticked);
-                    post.IsSticked = isSticked;
-                }
-
-                if (post.MetaInfo.ContainsKey("is-trending"))
-                {
-                    bool.TryParse(post.MetaInfo["is-trending"].FirstOrDefault(), out var isTrending);
-
-                    post.IsTrending = isTrending;
-                }
-
-                post.Template = "article";
-            }
-        }
-    }
-}
