--- VirtoCommerce.Storefront/Domain/StaticContent/ContentParsers/LangParser.cs
@@ -1,44 +1,0 @@
-using System.IO;
-using System.Linq;
-using VirtoCommerce.Storefront.Model;
-using VirtoCommerce.Storefront.Model.StaticContent;
-
-namespace VirtoCommerce.Storefront.Domain
-{
-    public partial class StaticContentItemBuilder
-    {
-        private class LangParser : IContentItemParser
-        {
-            public bool Suit(ContentItem item)
-            {
-                return true;
-            }
-
-            public void Parse(string path, string content, ContentItem item)
-            {
-                // each content file  has a name pattern {name}.{language?}.{ext}
-                var parts = Path.GetFileName(path)?.Split('.');
-
-                if (parts?.Length == 3)
-                {
-                    try
-                    {
-                        item.Language = new Language(parts[1]);
-                    }
-                    catch
-                    {
-                        // ignored
-                    }
-                }
-                if (item.Language == null)
-                {
-                    item.Language = Language.InvariantLanguage;
-                }
-                if (item.Name == null)
-                {
-                    item.Name = parts?.FirstOrDefault();
-                }
-            }
-        }
-    }
-}
