+++ VirtoCommerce.Storefront/Domain/StaticContent/ContentVisitors/PageMetadataVisitor.cs
@@ -6,25 +6,24 @@ using VirtoCommerce.Storefront.Model.StaticContent;
 
 namespace VirtoCommerce.Storefront.Domain
 {
-    public class PageBuilderContentLoader : IStaticContentLoader
+    internal class PageMetadataVisitor : IContentItemVisitor
     {
-        public string PrepareContent(string content)
+        public bool Suit(ContentItem item)
         {
-            return content;
+            return item.FileName.EndsWith(".page");
         }
 
-        public void ReadMetaData(string content, IDictionary<string, IEnumerable<string>> metadata)
+        public ContentItem Parse(string path, string content, ContentItem item)
         {
             var page = JsonConvert.DeserializeObject<JArray>(content);
             var settings = page.FirstOrDefault(x => (x as JObject)?.GetValue("type")?.Value<string>() == "settings");
             var items = settings.AsJEnumerable();
-
-            foreach (JProperty item in items)
+            item.MetaInfo = new Dictionary<string, IEnumerable<string>>();
+            foreach (JProperty prop in items)
             {
-                metadata.Add(item.Name, new List<string> { item.Value.Value<string>() });
+                item.MetaInfo.Add(prop.Name, new List<string> { prop.Value.Value<string>() });
             }
-
-            metadata.Add("template", new List<string> { "json-page" });
+            return item;
         }
     }
 }
